refactor: exporter structure
Major refactor of the exporter structure. The result is a renaming of
the "collector" concept to an "exporter" concept. The exporter now
accepts a context and is able to handle graceful shutdown via any method
of cancelation (we're currently using OS signals). Besides that, the
main function is quite a lot cleaner due to the moving of initialization
logic to other functions.

Signed-off-by: Luís Simas <luissimas@protonmail.com>